# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GoodHealth API
  description: |
    REST API for the GoodHealth fitness tracking application.

    ## Authentication
    All endpoints require authentication via Supabase Auth.
    Include the session token in the `Authorization` header as a Bearer token.

    ## Base URL
    - Development: `http://localhost:3000/api`
    - Production: `https://goodhealth-three.vercel.app/api`

  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/saurabhpro/goodhealth
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://goodhealth-three.vercel.app/api
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /workout-plans/generate:
    post:
      tags:
        - Workout Plans
      summary: Generate AI workout plan
      description: |
        Generates a complete AI-powered workout plan based on user goals.
        This endpoint:
        - Analyzes the user's goal and workout history
        - Selects appropriate workout templates
        - Creates a multi-week schedule with intelligent rotation
        - Applies progressive overload
        - Creates all database records (plan + sessions)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - goalId
                - weeksCount
              properties:
                goalId:
                  type: string
                  format: uuid
                  description: ID of the goal to base the plan on
                  example: "123e4567-e89b-12d3-a456-426614174000"
                weeksCount:
                  type: integer
                  minimum: 1
                  maximum: 12
                  description: Number of weeks for the plan
                  example: 4
                workoutsPerWeek:
                  type: integer
                  minimum: 3
                  maximum: 7
                  nullable: true
                  description: Optional override for workouts per week (calculated from goal if not provided)
                  example: 5
                avgDuration:
                  type: integer
                  nullable: true
                  description: Optional override for average workout duration in minutes
                  example: 60
                startDate:
                  type: string
                  format: date-time
                  nullable: true
                  description: Optional start date (defaults to today)
                name:
                  type: string
                  nullable: true
                  description: Optional custom plan name
                description:
                  type: string
                  nullable: true
                  description: Optional custom plan description
                preferences:
                  type: object
                  nullable: true
                  properties:
                    preferredDays:
                      type: array
                      items:
                        type: integer
                        minimum: 0
                        maximum: 6
                      description: Preferred workout days (0=Sunday, 6=Saturday)
                      example: [1, 3, 5]
                    avoidDays:
                      type: array
                      items:
                        type: integer
                        minimum: 0
                        maximum: 6
                      description: Days to avoid for workouts
                    focusAreas:
                      type: array
                      items:
                        type: string
                      description: Specific focus areas
                      example: ["upper_body", "cardio"]
      responses:
        '200':
          description: Plan generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  planId:
                    type: string
                    format: uuid
                  plan:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                      goal_type:
                        type: string
                        enum: [weight_loss, muscle_building, endurance, general_fitness]
                      weeks_duration:
                        type: integer
                      workouts_per_week:
                        type: integer
                      total_sessions:
                        type: integer
                      status:
                        type: string
                  schedule:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        $ref: '#/components/schemas/WorkoutPlanSession'
                    description: Weekly schedule with keys like 'week1', 'week2', etc.
                  summary:
                    type: object
                    properties:
                      totalWorkouts:
                        type: integer
                      avgWorkoutsPerWeek:
                        type: number
                        format: float
                      cardioSessions:
                        type: integer
                      strengthSessions:
                        type: integer
                      estimatedTotalTime:
                        type: integer
                        description: Total estimated time in minutes
        '400':
          description: Bad request (validation error)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Invalid weeks count: must be between 1 and 12"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Goal not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Goal not found"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase session token
  schemas:
...
    WorkoutPlanSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        week_number:
          type: integer
          minimum: 1
          example: 1
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: Day of week (0=Sunday, 6=Saturday)
          example: 1
        day_name:
          type: string
          enum: [Sunday, Monday, Tuesday, Wednesday, Thursday, Friday, Saturday]
          example: "Monday"
        session_order:
          type: integer
          minimum: 1
          default: 1
        workout_template_id:
          type: string
          format: uuid
          nullable: true
        workout_name:
          type: string
          example: "Upper Body Strength"
        workout_type:
          type: string
          enum: [strength, cardio, rest, active_recovery, mixed]
          example: "strength"
        estimated_duration:
          type: integer
          nullable: true
          description: Duration in minutes
          example: 60
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/Exercise'
        muscle_groups:
          type: array
          items:
            type: string
          example: ["chest", "triceps", "shoulders"]
        intensity_level:
          type: string
          enum: [low, moderate, high, max]
          nullable: true
          example: "high"
        status:
          type: string
          enum: [scheduled, completed, skipped]
          example: "scheduled"
        completed_workout_id:
          type: string
          format: uuid
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        skipped_at:
          type: string
          format: date-time
          nullable: true
        skip_reason:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Exercise:
      type: object
      properties:
        name:
          type: string
          example: "Bench Press"
        exercise_type:
          type: string
          enum: [strength, cardio]
          example: "strength"
        sets:
          type: integer
          nullable: true
          example: 3
        reps:
          type: integer
          nullable: true
          example: 10
        weight:
          type: number
          format: float
          nullable: true
          example: 60.0
        weight_unit:
          type: string
          nullable: true
          example: "kg"
        duration:
          type: integer
          nullable: true
          description: Duration in seconds (for cardio)
        distance:
          type: number
          format: float
          nullable: true
          description: Distance (for cardio)
        distance_unit:
          type: string
          nullable: true
          example: "km"
        notes:
          type: string
          nullable: true
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Not authenticated"
...